pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "plotly>=6.0.1",
    "requests>=2.32.3",
    "streamlit>=1.44.0",
]
